requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn==0.22.0
python-multipart  # for handling file uploads
gunicorn>=21.2.0  # for production server

# Image Processing
Pillow>=10.1.0
basicsr>=1.4.2
realesrgan>=0.3.0

# Excel Generation
openpyxl>=3.1.2
pandas>=2.1.3

# AI/ML
openai>=1.3.0
replicate>=0.20.0
google-genai>=0.6.0  # Google Gemini API

# Data Validation
pydantic>=2.5.0
pydantic-settings==2.3.4  # for settings management
requests

# Utils
python-jose[cryptography]>=3.3.0  # for JWT tokens
passlib[bcrypt]>=1.7.4  # for password hashing
python-dotenv==1.0.1  # for environment variables
aiofiles>=23.2.1  # for async file operations 
google-cloud-storage>=2.10.0  # Google Cloud Storage SDK
pandas
openpyxl